# docs/swagger.yaml
basePath: /app/history/api/
definitions:
  api.ReplyData-array_influx_Point:
    properties:
      data:
        items:
          $ref: '#/definitions/influx.Point'
        type: array
      error:
        type: string
    type: object
  influx.Point:
    properties:
      time:
        type: string
      value: {}
    type: object
info:
  contact: {}
  description: API文档
  title: 历史数据库接口文档
  version: 1.0 版本
paths:
  /query/{pid}/{id}/{name}:
    get:
      description: 查询历史数据
      parameters:
      - description: 产品ID
        in: path
        name: pid
        required: true
        type: string
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      - description: 变量名称
        in: path
        name: name
        required: true
        type: string
      - description: 起始时间
        in: query
        name: start
        type: string
      - description: 结束时间
        in: query
        name: end
        type: string
      - description: 窗口时间
        in: query
        name: window
        type: string
      - description: 算法
        in: query
        name: fn
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-array_influx_Point'
      summary: 查询历史数据
      tags:
      - query
  /query/{pid}/{id}/{name}/export:
    get:
      description: 导出历史数据
      parameters:
      - description: 产品ID
        in: path
        name: pid
        required: true
        type: string
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      - description: 变量名称
        in: path
        name: name
        required: true
        type: string
      - description: 起始时间
        in: query
        name: start
        type: string
      - description: 结束时间
        in: query
        name: end
        type: string
      - description: 窗口时间
        in: query
        name: window
        type: string
      - description: 算法
        in: query
        name: fn
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-array_influx_Point'
      summary: 导出历史数据
      tags:
      - query
  /rate/{pid}/{id}/{name}:
    get:
      description: 查询历史数据增长
      parameters:
      - description: 产品ID
        in: path
        name: pid
        required: true
        type: string
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      - description: 变量名称
        in: path
        name: name
        required: true
        type: string
      - description: 起始时间
        in: query
        name: start
        type: string
      - description: 结束时间
        in: query
        name: end
        type: string
      - description: 窗口时间
        in: query
        name: window
        type: string
      - description: 算法
        in: query
        name: fn
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-array_influx_Point'
      summary: 查询历史数据增长
      tags:
      - rate
  /rate/{pid}/{id}/{name}/export:
    get:
      description: 导出历史数据增长
      parameters:
      - description: 产品ID
        in: path
        name: pid
        required: true
        type: string
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      - description: 变量名称
        in: path
        name: name
        required: true
        type: string
      - description: 起始时间
        in: query
        name: start
        type: string
      - description: 结束时间
        in: query
        name: end
        type: string
      - description: 窗口时间
        in: query
        name: window
        type: string
      - description: 算法
        in: query
        name: fn
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-array_influx_Point'
      summary: 导出历史数据增长
      tags:
      - rate
swagger: "2.0"